requirements.txt
flask==2.3.3
pymongo==4.6.3
qrcode==7.4.2
pillow==10.4.0